refactor(api): move stage-repository.js to evaluation scope
