+++ api/lib/domain/usecases/get-user-campaign-assessment-result.js
@@ -2,7 +2,7 @@
 /* eslint-disable import/no-restricted-paths */
 import { NoCampaignParticipationForUserAndCampaign, NotFoundError } from '../errors.js';
 import * as defaultCompareStageAndAcquiredStagesService from '../services/stages/stage-and-stage-acquisition-comparison-service.js';
-import * as defaultStageRepository from '../../infrastructure/repositories/stage-repository.js';
+import * as defaultStageRepository from '../../../src/evaluation/infrastructure/repositories/stage-repository.js';
 import * as defaultStageAcquisitionRepository from '../../infrastructure/repositories/stage-acquisition-repository.js';
 import * as defaultParticipantResultRepository from '../../infrastructure/repositories/participant-result-repository.js';
