+++ api/lib/domain/usecases/stages/handle-stage-acquisition.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line eslint-comments/disable-enable-pair
 /* eslint-disable import/no-restricted-paths */
 import * as defaultSkillRepository from '../../../infrastructure/repositories/skill-repository.js';
-import * as defaultStageRepository from '../../../infrastructure/repositories/stage-repository.js';
+import * as defaultStageRepository from '../../../../src/evaluation/infrastructure/repositories/stage-repository.js';
 import * as defaultCampaignRepository from '../../../infrastructure/repositories/campaign-repository.js';
 import * as defaultStageAcquisitionRepository from '../../../infrastructure/repositories/stage-acquisition-repository.js';
 import * as defaultCampaignParticipationRepository from '../../../infrastructure/repositories/campaign-participation-repository.js';
