+++ api/src/evaluation/domain/usecases/index.js
@@ -8,6 +8,7 @@ import * as assessmentRepository from '../../../shared/infrastructure/repositori
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
 import * as competenceEvaluationRepository from '../../infrastructure/repositories/competence-evaluation-repository.js';
 import * as stageCollectionForTargetProfileRepository from '../../infrastructure/repositories/stage-collection-repository.js';
+import * as stageRepository from '../../infrastructure/repositories/stage-repository.js';
 import * as feedbackRepository from '../../infrastructure/repositories/feedback-repository.js';
 import * as stageRepository from '../../../../lib/infrastructure/repositories/stage-repository.js';
 import * as targetProfileForAdminRepository from '../../../../lib/infrastructure/repositories/target-profile-for-admin-repository.js';
